Refactored Service config stuff.
